+++ csharp/csharp_wrapper/MapInt64Int64.cs
@@ -11,7 +11,7 @@
 namespace TileDB {
 
 public class MapInt64Int64 : global::System.IDisposable 
-    , global::System.Collections.Generic.IDictionary<int, int>
+    , global::System.Collections.Generic.IDictionary<long, long>
  {
   private global::System.Runtime.InteropServices.HandleRef swigCPtr;
   protected bool swigCMemOwn;
@@ -47,7 +47,7 @@ public class MapInt64Int64 : global::System.IDisposable
   }
 
 
-  public int this[int key] {
+  public long this[long key] {
     get {
       return getitem(key);
     }
@@ -57,12 +57,12 @@ public class MapInt64Int64 : global::System.IDisposable
     }
   }
 
-  public bool TryGetValue(int key, out int value) {
+  public bool TryGetValue(long key, out long value) {
     if (this.ContainsKey(key)) {
       value = this[key];
       return true;
     }
-    value = default(int);
+    value = default(long);
     return false;
   }
 
@@ -78,9 +78,9 @@ public class MapInt64Int64 : global::System.IDisposable
     }
   }
 
-  public global::System.Collections.Generic.ICollection<int> Keys {
+  public global::System.Collections.Generic.ICollection<long> Keys {
     get {
-      global::System.Collections.Generic.ICollection<int> keys = new global::System.Collections.Generic.List<int>();
+      global::System.Collections.Generic.ICollection<long> keys = new global::System.Collections.Generic.List<long>();
       int size = this.Count;
       if (size > 0) {
         global::System.IntPtr iter = create_iterator_begin();
@@ -93,21 +93,21 @@ public class MapInt64Int64 : global::System.IDisposable
     }
   }
 
-  public global::System.Collections.Generic.ICollection<int> Values {
+  public global::System.Collections.Generic.ICollection<long> Values {
     get {
-      global::System.Collections.Generic.ICollection<int> vals = new global::System.Collections.Generic.List<int>();
-      foreach (global::System.Collections.Generic.KeyValuePair<int, int> pair in this) {
+      global::System.Collections.Generic.ICollection<long> vals = new global::System.Collections.Generic.List<long>();
+      foreach (global::System.Collections.Generic.KeyValuePair<long, long> pair in this) {
         vals.Add(pair.Value);
       }
       return vals;
     }
   }
 
-  public void Add(global::System.Collections.Generic.KeyValuePair<int, int> item) {
+  public void Add(global::System.Collections.Generic.KeyValuePair<long, long> item) {
     Add(item.Key, item.Value);
   }
 
-  public bool Remove(global::System.Collections.Generic.KeyValuePair<int, int> item) {
+  public bool Remove(global::System.Collections.Generic.KeyValuePair<long, long> item) {
     if (Contains(item)) {
       return Remove(item.Key);
     } else {
@@ -115,7 +115,7 @@ public class MapInt64Int64 : global::System.IDisposable
     }
   }
 
-  public bool Contains(global::System.Collections.Generic.KeyValuePair<int, int> item) {
+  public bool Contains(global::System.Collections.Generic.KeyValuePair<long, long> item) {
     if (this[item.Key] == item.Value) {
       return true;
     } else {
@@ -123,11 +123,11 @@ public class MapInt64Int64 : global::System.IDisposable
     }
   }
 
-  public void CopyTo(global::System.Collections.Generic.KeyValuePair<int, int>[] array) {
+  public void CopyTo(global::System.Collections.Generic.KeyValuePair<long, long>[] array) {
     CopyTo(array, 0);
   }
 
-  public void CopyTo(global::System.Collections.Generic.KeyValuePair<int, int>[] array, int arrayIndex) {
+  public void CopyTo(global::System.Collections.Generic.KeyValuePair<long, long>[] array, int arrayIndex) {
     if (array == null)
       throw new global::System.ArgumentNullException("array");
     if (arrayIndex < 0)
@@ -137,14 +137,14 @@ public class MapInt64Int64 : global::System.IDisposable
     if (arrayIndex+this.Count > array.Length)
       throw new global::System.ArgumentException("Number of elements to copy is too large.");
 
-    global::System.Collections.Generic.IList<int> keyList = new global::System.Collections.Generic.List<int>(this.Keys);
+    global::System.Collections.Generic.IList<long> keyList = new global::System.Collections.Generic.List<long>(this.Keys);
     for (int i = 0; i < keyList.Count; i++) {
-      int currentKey = keyList[i];
-      array.SetValue(new global::System.Collections.Generic.KeyValuePair<int, int>(currentKey, this[currentKey]), arrayIndex+i);
+      long currentKey = keyList[i];
+      array.SetValue(new global::System.Collections.Generic.KeyValuePair<long, long>(currentKey, this[currentKey]), arrayIndex+i);
     }
   }
 
-  global::System.Collections.Generic.IEnumerator<global::System.Collections.Generic.KeyValuePair<int, int>> global::System.Collections.Generic.IEnumerable<global::System.Collections.Generic.KeyValuePair<int, int>>.GetEnumerator() {
+  global::System.Collections.Generic.IEnumerator<global::System.Collections.Generic.KeyValuePair<long, long>> global::System.Collections.Generic.IEnumerable<global::System.Collections.Generic.KeyValuePair<long, long>>.GetEnumerator() {
     return new MapInt64Int64Enumerator(this);
   }
 
@@ -162,24 +162,24 @@ public class MapInt64Int64 : global::System.IDisposable
   /// collection but not when one of the elements of the collection is modified as it is a bit
   /// tricky to detect unmanaged code that modifies the collection under our feet.
   public sealed class MapInt64Int64Enumerator : global::System.Collections.IEnumerator,
-      global::System.Collections.Generic.IEnumerator<global::System.Collections.Generic.KeyValuePair<int, int>>
+      global::System.Collections.Generic.IEnumerator<global::System.Collections.Generic.KeyValuePair<long, long>>
   {
     private MapInt64Int64 collectionRef;
-    private global::System.Collections.Generic.IList<int> keyCollection;
+    private global::System.Collections.Generic.IList<long> keyCollection;
     private int currentIndex;
     private object currentObject;
     private int currentSize;
 
     public MapInt64Int64Enumerator(MapInt64Int64 collection) {
       collectionRef = collection;
-      keyCollection = new global::System.Collections.Generic.List<int>(collection.Keys);
+      keyCollection = new global::System.Collections.Generic.List<long>(collection.Keys);
       currentIndex = -1;
       currentObject = null;
       currentSize = collectionRef.Count;
     }
 
     // Type-safe iterator Current
-    public global::System.Collections.Generic.KeyValuePair<int, int> Current {
+    public global::System.Collections.Generic.KeyValuePair<long, long> Current {
       get {
         if (currentIndex == -1)
           throw new global::System.InvalidOperationException("Enumeration not started.");
@@ -187,7 +187,7 @@ public class MapInt64Int64 : global::System.IDisposable
           throw new global::System.InvalidOperationException("Enumeration finished.");
         if (currentObject == null)
           throw new global::System.InvalidOperationException("Collection modified.");
-        return (global::System.Collections.Generic.KeyValuePair<int, int>)currentObject;
+        return (global::System.Collections.Generic.KeyValuePair<long, long>)currentObject;
       }
     }
 
@@ -203,8 +203,8 @@ public class MapInt64Int64 : global::System.IDisposable
       bool moveOkay = (currentIndex+1 < size) && (size == currentSize);
       if (moveOkay) {
         currentIndex++;
-        int currentKey = keyCollection[currentIndex];
-        currentObject = new global::System.Collections.Generic.KeyValuePair<int, int>(currentKey, collectionRef[currentKey]);
+        long currentKey = keyCollection[currentIndex];
+        currentObject = new global::System.Collections.Generic.KeyValuePair<long, long>(currentKey, collectionRef[currentKey]);
       } else {
         currentObject = null;
       }
@@ -247,27 +247,27 @@ public class MapInt64Int64 : global::System.IDisposable
     tiledbcsPINVOKE.MapInt64Int64_Clear(swigCPtr);
   }
 
-  private int getitem(int key) {
-    int ret = tiledbcsPINVOKE.MapInt64Int64_getitem(swigCPtr, key);
+  private long getitem(long key) {
+    long ret = tiledbcsPINVOKE.MapInt64Int64_getitem(swigCPtr, key);
     if (tiledbcsPINVOKE.SWIGPendingException.Pending) throw tiledbcsPINVOKE.SWIGPendingException.Retrieve();
     return ret;
   }
 
-  private void setitem(int key, int x) {
+  private void setitem(long key, long x) {
     tiledbcsPINVOKE.MapInt64Int64_setitem(swigCPtr, key, x);
   }
 
-  public bool ContainsKey(int key) {
+  public bool ContainsKey(long key) {
     bool ret = tiledbcsPINVOKE.MapInt64Int64_ContainsKey(swigCPtr, key);
     return ret;
   }
 
-  public void Add(int key, int value) {
+  public void Add(long key, long value) {
     tiledbcsPINVOKE.MapInt64Int64_Add(swigCPtr, key, value);
     if (tiledbcsPINVOKE.SWIGPendingException.Pending) throw tiledbcsPINVOKE.SWIGPendingException.Retrieve();
   }
 
-  public bool Remove(int key) {
+  public bool Remove(long key) {
     bool ret = tiledbcsPINVOKE.MapInt64Int64_Remove(swigCPtr, key);
     return ret;
   }
@@ -277,8 +277,8 @@ public class MapInt64Int64 : global::System.IDisposable
     return ret;
   }
 
-  private int get_next_key(global::System.IntPtr swigiterator) {
-    int ret = tiledbcsPINVOKE.MapInt64Int64_get_next_key(swigCPtr, swigiterator);
+  private long get_next_key(global::System.IntPtr swigiterator) {
+    long ret = tiledbcsPINVOKE.MapInt64Int64_get_next_key(swigCPtr, swigiterator);
     return ret;
   }
